chore(contracts): extract and share chunkEntriesBatch
