+++ packages/serverless-contracts/src/contracts/SQS/utils/index.ts
@@ -5,4 +5,3 @@ export * from './serializeMessageAttributes';
 export * from './messagesValidation';
 export * from './serializeMessage';
 export * from './delays';
-export * from './chunckSQSMessagesBatch';
