+++ packages/serverless-contracts/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './fillPathTemplate';
 export * from './isDefined';
 export * from './typeAssert';
+export * from './chunkEntriesBatch';
